// src/app/api/core/model/delivery/transaction/tracking/interfaces/transaction-tracking-action.interface.ts
import { TransactionTrackingActionDetail } from './transaction-tracking-action-detail.interface';
import { TransactionTrackingState } from './transaction-tracking-state.interface';
import { TransactionTrackingStyle } from './transaction-tracking-style.interface';

export interface TransactionTrackingAction {
  action: TransactionTrackingActionDetail;
  state: TransactionTrackingState;
  style: TransactionTrackingStyle;
  title: string;
}